feat(libsy): move stage router to libsy

feat(libsy): bridge back to components for stage_router
