+++ crates/libsy/src/algorithms/util.rs
@@ -1,8 +1,10 @@
 // SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod affinity;
-pub mod subagent;
+mod affinity;
+pub(crate) mod stage_router;
+mod subagent;
+pub(crate) mod tool_signals;
 
 pub use affinity::AffinityRouter;
 pub use subagent::SubagentOverride;
