--- crates/libsy/src/signal.rs
@@ -1,5 +1,0 @@
-// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod tool_signals;
-pub use tool_signals::{ToolSignals, DEFAULT_RECENT_WINDOW};
